Config related issues resolved

+++ HumanitarianAssistance.Application/Accounting/Models/SPVoucherSummaryReportModel.cs
@@ -2,7 +2,7 @@ namespace HumanitarianAssistance.Application.Accounting.Models
 {
     public class SPVoucherSummaryReportModel
     {
-        public int VoucherNo { get; set; }
+        public long VoucherNo { get; set; }
         public string VoucherCode { get; set; }
         public string VoucherDescription { get; set; }
         public string VoucherDate { get; set; }
